# swagger/api.yaml
openapi: "3.0.0"
info:
  title: BG Library
  version: "1.0"
servers:
  - url: ${API_URL}
    description: Dynamic Server
paths:
...
  /api/v1/library/patron:
    post:
      tags:
        - patrons
      operationId: addPatron
      summary: Add a patron
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatronRequest'
      responses:
        '201':
          description: Patron added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patron'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/patrons:
    get:
      tags:
        - patrons
      operationId: listPatrons
      summary: Get list of all patrons
      parameters:
        - in: query
          name: name
          schema:
            type: string
          description: Filter patrons by name (optional)
      responses:
        '200':
          description: List of games with checkout status successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatronList'
    post:
      tags:
        - patrons
      operationId: bulkAddPatrons
      summary: Add multiple patrons
      description: |
        Add multiple patrons by uploading a CSV file.
        This call is all or nothing. If any patron fails to be added, none will be added.
        The CSV file should contain a list of patrons names, one per line, in the following format: 
        "John Smith"\n"Jane Doe"\n"Alice Baker"
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              format: byte
              description: Base64 encoded CSV of patrons to add
      responses:
        '201':
          description: Patrons added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkAddResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/patron/id/{patronId}:
    get:
      tags:
        - patrons
      operationId: getPatron
      summary: Get a patron
      parameters:
        - in: path
          name: patronId
          schema:
            type: string
            format: uuid
          required: true
          description: Patron ID
      responses:
        '200':
          description: Patron found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patron'
        '404':
          description: Patron not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - patrons
      operationId: updatePatron
      summary: Update an existing patron
      parameters:
        - in: path
          name: patronId
          schema:
            type: string
            format: uuid
          required: true
          description: Game ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatronRequest'
      responses:
        '204':
          description: Patron successfully updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patron not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - patrons
      operationId: deletePatron
      summary: Soft delete a patron
      parameters:
        - in: path
          name: patronId
          schema:
            type: string
            format: uuid
          required: true
          description: Patron ID
      responses:
        '204':
          description: Patron deleted successfully
        '404':
          description: Patron not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/library/patron/barcode/{patronBarcode}:
    get:
      tags:
        - patrons
        - barcode
      operationId: getPatronByBarcode
      summary: Get a patron from a barcode
      parameters:
        - in: path
          name: patronBarcode
          schema:
            type: string
          required: true
          description: Patron Barcode
      responses:
        '200':
          description: Patron found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patron'
        '404':
          description: Patron not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Patron:
      type: object
      required:
        - patronId
        - name
      properties:
        patronId:
          type: string
          format: uuid
          description: Patron ID
          example: "e1e07b70-c9f3-442f-8225-fe468c3ae705"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the patron
          example: "John Smith"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the patron
          example: "9780307455925"
      description: A library patron
    CreatePatronRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: The name of the patron
          example: "John Smith"
        barcode:
          type: string
          minLength: 1
          maxLength: 48
          description: The barcode of the patron
          example: "9780307455925"
      description: Request payload for creating a new patron
...
    PatronList:
      type: object
      required:
        - patrons
      properties:
        patrons:
          type: array
          items:
            $ref: '#/components/schemas/Patron'
          description: Array of patrons
      description: List of patrons with check out info
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
                - NOT_FOUND
                - CONFLICT
                - MALFORMED_REQUEST
              description: Machine readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Validation Error"
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
              description: Array of specific field errors. Empty if no specific errors
      description: Response when errors occur
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: The input field that caused the error
          example: "title"
        message:
          type: string
          description: Human-readable error message for this field
          example: "Title cannot exceed 100 characters"
      description: Details about specific errors
    BulkAddResponse:
      type: object
      required:
        - imported
      properties:
        imported:
          type: integer
          description: Number of records imported successfully
          example: 10
      description: Response for bulk add operations
...